Use official repository for sketch compilation CI action
The official dedicated repository for the Arduino sketch compilation GitHub Actions action is now
arduino/compile-sketches. Since the time the action was moved, there was a breaking change to the default sketches
report file name. Rather than continuing to rely on the default value, a specific file name is now defined in the
workflow.