Use custom matrix job names in sketch compilation CI job
The "Compile Examples" GitHub Actions workflow generates a matrix job for each board. The default job name is generated
from the job's matrix object. This contains the complete board data, which results in a long and somewhat cryptic job
name that can make the workflow run more difficult to interpret.

The only necessary information is the FQBN. A custom job name allows for only using this information in the job name.